requirements.txt
flask
openai
langchain
python-dotenv
databricks-sql-connector
langchain-community
langchain-openai